refactor: repositoty 계층 삭제
