--- src/main/java/roomescape/domain/member/repository/MemberRepository.java
@@ -1,17 +1,0 @@
-package roomescape.domain.member.repository;
-
-import java.util.List;
-import roomescape.domain.member.model.Member;
-
-public interface MemberRepository {
-
-    void save(Member member);
-
-    Member findByEmail(String email);
-
-    Member findByEmailAndPassword(String email, String password);
-
-    Member findById(Long memberId);
-
-    List<Member> findAll();
-}
