+++ src/main/java/roomescape/domain/reservation/service/ReservationService.java
@@ -5,61 +5,69 @@ import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.currentdate.CurrentDateTime;
+import roomescape.domain.member.dao.MemberDao;
 import roomescape.domain.member.model.Member;
-import roomescape.domain.member.repository.MemberRepository;
-import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationDate;
+import roomescape.domain.reservation.dao.ReservationDao;
 import roomescape.domain.reservation.dto.request.AdminReservationRequestDto;
 import roomescape.domain.reservation.dto.request.ReservationRequestDto;
 import roomescape.domain.reservation.dto.response.AdminReservationResponse;
 import roomescape.domain.reservation.dto.response.ReservationResponseDto;
-import roomescape.domain.reservation.repository.ReservationRepository;
-import roomescape.domain.reservationtime.ReservationTime;
-import roomescape.domain.reservationtime.repository.ReservationTimeRepository;
+import roomescape.domain.reservation.model.Reservation;
+import roomescape.domain.reservation.model.ReservationDate;
+import roomescape.domain.reservationtime.dao.ReservationTimeDao;
+import roomescape.domain.reservationtime.model.ReservationTime;
+import roomescape.domain.theme.dao.ThemeDao;
 import roomescape.domain.theme.model.Theme;
-import roomescape.domain.theme.repository.ThemeRepository;
 import roomescape.global.exception.InvalidReservationException;
 
 @Service
 public class ReservationService {
 
-    private final MemberRepository memberRepository;
-    private final ReservationRepository reservationRepository;
-    private final ReservationTimeRepository reservationTimeRepository;
-    private final ThemeRepository themeRepository;
+    private final MemberDao memberDao;
+    private final ReservationDao reservationDao;
+    private final ReservationTimeDao reservationTimeDao;
+    private final ThemeDao themeDao;
     private final CurrentDateTime currentDateTime;
 
     public ReservationService(
-        MemberRepository memberRepository,
-        ReservationRepository reservationRepository,
-        ReservationTimeRepository reservationTimeRepository,
-        ThemeRepository themeRepository,
+        MemberDao memberDao,
+        ReservationDao reservationDao,
+        ReservationTimeDao reservationTimeDao,
+        ThemeDao themeDao,
         CurrentDateTime currentDateTime) {
-        this.memberRepository = memberRepository;
-        this.reservationRepository = reservationRepository;
-        this.reservationTimeRepository = reservationTimeRepository;
-        this.themeRepository = themeRepository;
+        this.memberDao = memberDao;
+        this.reservationDao = reservationDao;
+        this.reservationTimeDao = reservationTimeDao;
+        this.themeDao = themeDao;
         this.currentDateTime = currentDateTime;
     }
 
     public List<ReservationResponseDto> getAllReservations() {
-        return reservationRepository.findAll().stream()
+        return reservationDao.findAll().stream()
             .map(ReservationResponseDto::from)
             .toList();
     }
 
     public void deleteReservation(Long id) {
-        reservationRepository.delete(id);
+        findReservationBy(id);
+        reservationDao.delete(id);
     }
 
-    public ReservationResponseDto saveReservationOfMember(ReservationRequestDto request,
+    public void findReservationBy(Long id) {
+        reservationDao.findById(id).
+            orElseThrow(() -> new IllegalArgumentException("존재하지 않는 예약 ID 입니다."));
+    }
+
+    public ReservationResponseDto saveReservationOfMember(
+        ReservationRequestDto request,
         long memberId) {
         Reservation reservation = saveReservation(
             request.date(), memberId, request.timeId(), request.themeId());
         return ReservationResponseDto.from(reservation);
     }
 
-    public AdminReservationResponse saveReservationOfAdmin(AdminReservationRequestDto request,
+    public AdminReservationResponse saveReservationOfAdmin(
+        AdminReservationRequestDto request,
         long memberId) {
         Reservation reservation = saveReservation(
             request.date(), memberId, request.timeId(), request.themeId());
@@ -67,41 +75,52 @@ public class ReservationService {
     }
 
     private Reservation saveReservation(String date, long memberId, long timeId, long themeId) {
-        Member member = memberRepository.findById(memberId);
+        Member member = findMemberBy(memberId);
         Reservation reservation = createReservation(member, date, timeId, themeId);
         validateDateTimeAndSaveReservation(reservation, themeId);
         return reservation;
+    }
+
+    public Member findMemberBy(long memberId) {
+        return memberDao.findById(memberId)
+            .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
     }
 
     private Reservation createReservation(Member member, String date, long timeId, long themeId) {
         LocalDateTime currentDateTimeInfo = currentDateTime.get();
         ReservationDate reservationDate = new ReservationDate(LocalDate.parse(date));
         reservationDate.validateDate(currentDateTimeInfo.toLocalDate());
-        ReservationTime reservationTime = reservationTimeRepository.findById(timeId);
-        Theme theme = themeRepository.findById(themeId);
+        ReservationTime reservationTime = findReservationTimeBy(timeId);
+        Theme theme = findThemeBy(themeId);
         return new Reservation(member, reservationDate, reservationTime, theme);
+    }
+
+    private Theme findThemeBy(long themeId) {
+        return themeDao.findById(themeId)
+            .orElseThrow(() -> new IllegalArgumentException("해당 ID의 테마를 찾을 수 없습니다"));
+    }
+
+    private ReservationTime findReservationTimeBy(long timeId) {
+        return reservationTimeDao.findById(timeId)
+            .orElseThrow(() -> new IllegalArgumentException("해당 ID의 예약시간을 찾을 수 없습니다"));
     }
 
     private void validateDateTimeAndSaveReservation(Reservation reservation, long timeId) {
         reservation.validateDateTime(currentDateTime.get());
         validateAlreadyExistDateTime(reservation.getReservationDate(), timeId);
-        reservationRepository.save(reservation);
+        long savedId = reservationDao.save(reservation);
+        reservation.setId(savedId);
     }
 
     private void validateAlreadyExistDateTime(ReservationDate date, long timeId) {
-        if (reservationRepository.hasAnotherReservation(date, timeId)) {
+        if (reservationDao.existReservationOf(date, timeId)) {
             throw new InvalidReservationException("중복된 날짜와 시간을 예약할 수 없습니다.");
         }
     }
 
     public List<ReservationResponseDto> getAllReservationsOf(
-        String dateFrom,
-        String dateTo,
-        Long memberId,
-        Long themeId
-    ) {
-        List<Reservation> reservations = reservationRepository.findOf(
-            dateFrom, dateTo, memberId, themeId);
+        String dateFrom, String dateTo, Long memberId, Long themeId) {
+        List<Reservation> reservations = reservationDao.findOf(dateFrom, dateTo, memberId, themeId);
         return reservations.stream()
             .map(ReservationResponseDto::from)
             .toList();
