+++ src/test/java/roomescape/dao/JdbcReservationTimeDaoTest.java
@@ -11,9 +11,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.domain.reservationtime.dao.ReservationTimeDao;
 import roomescape.domain.reservationtime.dto.response.BookedReservationTimeResponseDto;
+import roomescape.domain.reservationtime.model.ReservationTime;
 
 @Sql(scripts = {"/test-schema.sql"})
 @SpringBootTest
