+++ src/test/java/roomescape/domain/ReservationDateTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.reservation.ReservationDate;
+import roomescape.domain.reservation.model.ReservationDate;
 
 class ReservationDateTest {
