+++ src/test/java/roomescape/service/MemberServiceTest.java
@@ -1,0 +1,5 @@
+package roomescape.service;
+
+public class MemberServiceTest {
+
+}
