+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.service;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -12,37 +12,35 @@ import roomescape.dao.FakeReservationDaoImpl;
 import roomescape.dao.FakeReservationTimeDaoImpl;
 import roomescape.dao.FakeThemeDaoImpl;
 import roomescape.domain.member.dao.MemberDao;
-import roomescape.domain.reservation.dao.ReservationDao;
-import roomescape.domain.reservationtime.dao.ReservationTimeDao;
-import roomescape.domain.theme.dao.ThemeDao;
 import roomescape.domain.member.model.Member;
-import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationDate;
-import roomescape.domain.reservationtime.ReservationTime;
+import roomescape.domain.reservation.dao.ReservationDao;
+import roomescape.domain.reservation.model.Reservation;
+import roomescape.domain.reservation.model.ReservationDate;
+import roomescape.domain.reservationtime.dao.ReservationTimeDao;
+import roomescape.domain.reservationtime.model.ReservationTime;
+import roomescape.domain.theme.dao.ThemeDao;
 import roomescape.domain.theme.model.Theme;
-import roomescape.domain.theme.repository.ThemeRepository;
+import roomescape.domain.theme.service.ThemeService;
 import roomescape.global.exception.InvalidReservationException;
-import roomescape.domain.theme.repository.ThemeRepositoryImpl;
 
-public class ThemeRepositoryTest {
+public class ThemeServiceTest {
 
     private final MemberDao memberDao = new FakeMemberDaoImpl();
     private final ThemeDao themeDao = new FakeThemeDaoImpl();
     private final ReservationTimeDao reservationTimeDao = new FakeReservationTimeDaoImpl();
     private final ReservationDao reservationDao = new FakeReservationDaoImpl();
-    private final ThemeRepository themeRepository = new ThemeRepositoryImpl(themeDao,
+    private final ThemeService themeService = new ThemeService(themeDao,
         reservationDao);
 
     @Test
     @DisplayName("존재하지 않는 테마ID를 가져오려고 할 경우, 예외가 발생해야 한다.")
     void not_exist_theme_id_get_then_throw_exception() {
-        assertThatThrownBy(() -> themeRepository.findById(999L))
+        assertThatThrownBy(() -> themeService.findById(999L))
             .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     void delete_already_use_other_theme_id_then_throw_exception() {
-
         //given
         Member member = Member.createUser("testMember", "a@email.com", "aaa");
         long savedMemberId = memberDao.save(member);
@@ -68,7 +66,7 @@ public class ThemeRepositoryTest {
 
         //when
         assertThatThrownBy(
-            () -> themeRepository.delete(savedThemeId))
+            () -> themeService.deleteTheme(savedThemeId))
             .isInstanceOf(InvalidReservationException.class)
             .hasMessage("이미 예약된 테마를 삭제할 수 없습니다.");
     }
@@ -79,9 +77,8 @@ public class ThemeRepositoryTest {
         Theme theme = new Theme("공포", "공포테마입니다", "http://aaa");
         long savedId = themeDao.save(theme);
         theme.setId(savedId);
-        assertThatCode(() -> themeRepository.findById(savedId)).doesNotThrowAnyException();
+        assertThatCode(() -> themeService.findById(savedId)).doesNotThrowAnyException();
     }
-
 
     @Test
     @DisplayName("존재하지 않는 테마ID를 삭제하려고 할 경우, 예외가 발생해야 한다.")
@@ -89,6 +86,6 @@ public class ThemeRepositoryTest {
         Theme theme = new Theme("공포", "공포테마입니다", "http://aaa");
         long savedId = themeDao.save(theme);
         theme.setId(savedId);
-        assertThatCode(() -> themeRepository.delete(savedId)).doesNotThrowAnyException();
+        assertThatCode(() -> themeService.deleteTheme(savedId)).doesNotThrowAnyException();
     }
 }
